+++ test/spec/Feature/Query/HtmlRawOutputSpec.hs
@@ -12,7 +12,8 @@ import SpecHelper (acceptHdrs)
 
 spec :: SpecWith ((), Application)
 spec = describe "When raw-media-types is set to \"text/html\"" $
-  it "can get raw output with Accept: text/html" $
+  it "can get raw output with Accept: text/html" $ do
+    pendingWith "TBD"
     request methodGet "/rpc/welcome.html" (acceptHdrs "text/html") ""
       `shouldRespondWith`
       [str|